rename subsystem and channel modules to base_subsystem and base_channel to avoid conflict when importing from lantz_core.
