+++ tests/test_subsytem.py
@@ -13,7 +13,7 @@ from __future__ import (division, unicode_literals, print_function,
                         absolute_import)
 
 from lantz_core.has_features import subsystem
-from lantz_core.subsystem import SubSystem
+from lantz_core.base_subsystem import SubSystem
 from .testing_tools import DummyParent
